+++ src/ezdxf/appsettings.py
@@ -7,6 +7,7 @@ from typing import TYPE_CHECKING
 from ezdxf.lldxf import const, validator
 from ezdxf.math import BoundingBox
 from ezdxf import bbox
+from ezdxf.enums import EndCaps, JoinStyle
 
 if TYPE_CHECKING:
     from ezdxf.document import Drawing
@@ -117,3 +118,13 @@ def show_lineweight(doc: Drawing, state=True) -> None:
     "thickness" (lineweight) if `state` is ``True``.
     """
     doc.header["$LWDISPLAY"] = int(state)
+
+
+def set_lineweight_display_style(
+    doc: Drawing, end_caps: EndCaps, join_style: JoinStyle
+) -> None:
+    """Set the style of end caps and joints for linear entities when displaying
+    line weights. These settings only affect objects created afterwards.
+    """
+    doc.header["$ENDCAPS"] = int(end_caps)
+    doc.header["$JOINSTYLE"] = int(join_style)